Move AgentUiState to viewmodels module
