+++ app-shared/src/commonTest/kotlin/org/pointyware/artes/agents/ui/NewAgentViewUnitTest.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
 import androidx.compose.ui.test.runComposeUiTest
-import org.pointyware.artes.agents.viewmodels.AgentUiState
+import org.pointyware.artes.viewmodels.AgentUiState
 import org.pointyware.artes.viewmodels.HostUiState
 import org.pointyware.artes.viewmodels.ModelUiState
 import kotlin.test.AfterTest
