+++ core-viewmodels/src/commonMain/kotlin/org/pointyware/artes/viewmodels/HostUiState.kt
@@ -5,6 +5,8 @@ import org.pointyware.artes.entities.Host
 /**
  * Complete information about a service to display in a detailed view.
  *
+ * TODO: rename to HostInfoUiState - or other after rename
+ *
  * @see [org.pointyware.artes.entities.Host]
  */
 data class HostUiState(
